Avoid defaults in the oracle-aware extractors.
