+++ include/tatami/base/Matrix.hpp
@@ -329,9 +329,7 @@ public:
      * @return A `FullDenseOracleAwareExtractor` object for dense access to full rows.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<FullDenseOracleAwareExtractor<Value_, Index_> > dense_row(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::FULL, false, Value_, Index_> >(this->dense_row(opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<FullDenseOracleAwareExtractor<Value_, Index_> > dense_row(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const = 0;
 
     /**
      * @param block_start Index of the column at the start of the block.
@@ -341,9 +339,7 @@ public:
      * @return A `BlockDenseOracleAwareExtractor` object for dense access to a block of each row.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<BlockDenseOracleAwareExtractor<Value_, Index_> > dense_row(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::BLOCK, false, Value_, Index_> >(this->dense_row(block_start, block_length, opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<BlockDenseOracleAwareExtractor<Value_, Index_> > dense_row(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const = 0;
 
     /**
      * @param indices Vector of sorted and unique column indices.
@@ -352,9 +348,7 @@ public:
      * @return A `IndexDenseOracleAwareExtractor` object for dense access to a indexed subset of each row.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<IndexDenseOracleAwareExtractor<Value_, Index_> > dense_row(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::INDEX, false, Value_, Index_> >(this->dense_row(std::move(indices), opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<IndexDenseOracleAwareExtractor<Value_, Index_> > dense_row(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const = 0;
 
     /**
      * @param oracle An oracle supplying iteration predictions.
@@ -362,9 +356,7 @@ public:
      * @return A `FullDenseOracleAwareExtractor` object for dense access to full columns.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<FullDenseOracleAwareExtractor<Value_, Index_> > dense_column(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::FULL, false, Value_, Index_> >(this->dense_column(opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<FullDenseOracleAwareExtractor<Value_, Index_> > dense_column(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const = 0;
 
     /**
      * @param block_start Index of the row at the start of the block.
@@ -374,9 +366,7 @@ public:
      * @return A `BlockDenseOracleAwareExtractor` object for dense access to a block of each column.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<BlockDenseOracleAwareExtractor<Value_, Index_> > dense_column(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::BLOCK, false, Value_, Index_> >(this->dense_column(block_start, block_length, opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<BlockDenseOracleAwareExtractor<Value_, Index_> > dense_column(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const = 0;
 
     /**
      * @param indices Vector of sorted and unique column indices.
@@ -385,13 +375,11 @@ public:
      * @return A `IndexDenseOracleAwareExtractor` object for dense access to a indexed subset of each column.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<IndexDenseOracleAwareExtractor<Value_, Index_> > dense_column(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::INDEX, false, Value_, Index_> >(this->dense_column(std::move(indices), opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<IndexDenseOracleAwareExtractor<Value_, Index_> > dense_column(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const = 0;
 
-    /***************************************
-     **** Sparse access virtual methods ****
-     ***************************************/
+    /****************************************************
+     **** Sparse oracle-aware access virtual methods ****
+     ****************************************************/
 public:
     /**
      * @param oracle An oracle supplying iteration predictions.
@@ -399,9 +387,7 @@ public:
      * @return A `FullSparseOracleAwareExtractor` object for sparse access to full rows.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<FullSparseOracleAwareExtractor<Value_, Index_> > sparse_row(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::FULL, true, Value_, Index_> >(this->sparse_row(opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<FullSparseOracleAwareExtractor<Value_, Index_> > sparse_row(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const = 0;
 
     /**
      * @param block_start Index of the column at the start of the block.
@@ -411,9 +397,7 @@ public:
      * @return A `BlockSparseOracleAwareExtractor` object for sparse access to a block of each row.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<BlockSparseOracleAwareExtractor<Value_, Index_> > sparse_row(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::BLOCK, true, Value_, Index_> >(this->sparse_row(block_start, block_length, opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<BlockSparseOracleAwareExtractor<Value_, Index_> > sparse_row(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const = 0;
 
     /**
      * @param indices Vector of sorted and unique column indices.
@@ -422,9 +406,7 @@ public:
      * @return A `IndexSparseOracleAwareExtractor` object for sparse access to a indexed subset of each row.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<IndexSparseOracleAwareExtractor<Value_, Index_> > sparse_row(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::INDEX, true, Value_, Index_> >(this->sparse_row(std::move(indices), opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<IndexSparseOracleAwareExtractor<Value_, Index_> > sparse_row(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const = 0;
 
     /**
      * @param oracle An oracle supplying iteration predictions.
@@ -432,9 +414,7 @@ public:
      * @return A `FullSparseOracleAwareExtractor` object for sparse access to full columns.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<FullSparseOracleAwareExtractor<Value_, Index_> > sparse_column(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::FULL, true, Value_, Index_> >(this->sparse_column(opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<FullSparseOracleAwareExtractor<Value_, Index_> > sparse_column(std::shared_ptr<Oracle<Index_> > oracle, const Options& opt) const = 0;
 
     /**
      * @param block_start Index of the row at the start of the block.
@@ -444,9 +424,7 @@ public:
      * @return A `BlockSparseOracleAwareExtractor` object for sparse access to a block of each column.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<BlockSparseOracleAwareExtractor<Value_, Index_> > sparse_column(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::BLOCK, true, Value_, Index_> >(this->sparse_column(block_start, block_length, opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<BlockSparseOracleAwareExtractor<Value_, Index_> > sparse_column(std::shared_ptr<Oracle<Index_> > oracle, Index_ block_start, Index_ block_length, const Options& opt) const = 0;
 
     /**
      * @param indices Vector of sorted and unique column indices.
@@ -455,9 +433,7 @@ public:
      * @return A `IndexSparseOracleAwareExtractor` object for sparse access to a indexed subset of each column.
      * This should not outlive the parent `Matrix` from which it was created.
      */
-    virtual std::unique_ptr<IndexSparseOracleAwareExtractor<Value_, Index_> > sparse_column(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const {
-        return std::make_unique<DummyOracleAwareExtractor<DimensionSelectionType::INDEX, true, Value_, Index_> >(this->sparse_column(std::move(indices), opt), std::move(oracle));
-    }
+    virtual std::unique_ptr<IndexSparseOracleAwareExtractor<Value_, Index_> > sparse_column(std::shared_ptr<Oracle<Index_> > oracle, std::vector<Index_> indices, const Options& opt) const = 0;
 
     /****************************************************
      **** Dense oracle-aware access default overload ****
@@ -523,9 +499,9 @@ public:
         return dense_column(std::move(oracle), std::move(indices), Options()); 
     }
 
-    /*****************************************
-     **** Sparse access default overloads ****
-     *****************************************/
+    /*****************************************************
+     **** Sparse oracle-aware access default overload ****
+     *****************************************************/
 public:
     /**
      * @param oracle An oracle supplying iteration predictions.
