+++ include/tatami/sparse/FragmentedSparseMatrix.hpp
@@ -6,6 +6,7 @@
 #include "primary_extraction.hpp"
 #include "SparseSecondaryExtractorCore.hpp"
 #include "../utils/ElementType.hpp"
+#include "../utils/OracleUnawareMatrix.hpp"
 
 #include <vector>
 #include <algorithm>
@@ -52,7 +53,7 @@ template<
     class ValueVectorStorage_ = std::vector<std::vector<Value_> >,
     class IndexVectorStorage_ = std::vector<std::vector<Index_> >
 >
-class FragmentedSparseMatrix : public Matrix<Value_, Index_> {
+class FragmentedSparseMatrix : public OracleUnawareMatrix<Value_, Index_> {
 public:
     /**
      * @param nr Number of rows.
@@ -171,10 +172,6 @@ private:
             } else {
                 return NULL;
             }
-        }
-
-        void set_oracle(std::unique_ptr<Oracle<Index_> >) {
-            return;
         }
 
     protected:
