+++ include/tatami/sparse/SemiCompressedSparseMatrix.hpp
@@ -6,6 +6,7 @@
 #include "utils.hpp"
 #include "SparseSecondaryExtractorCore.hpp"
 #include "../utils/ElementType.hpp"
+#include "../utils/OracleUnawareMatrix.hpp"
 
 #include <vector>
 #include <algorithm>
@@ -49,7 +50,7 @@ template<
     class IndexStorage_ = std::vector<Index_>, 
     class PointerStorage_ = std::vector<size_t> 
 >
-class SemiCompressedSparseMatrix : public Matrix<Value_, Index_> {
+class SemiCompressedSparseMatrix : public OracleUnawareMatrix<Value_, Index_> {
 public:
     /**
      * @param nr Number of rows.
@@ -171,10 +172,6 @@ private:
             } else {
                 return NULL;
             }
-        }
-
-        void set_oracle(std::unique_ptr<Oracle<Index_> >) {
-            return;
         }
 
     protected:
